File refactor
