+++ src/PhpDia/Dia/Xml/Attribute.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace PhpDia\Dia;
+namespace PhpDia\Dia\Xml;
+
+use PhpDia\Dia\RenderItem;
+use PhpDia\Dia\TemplateManager;
 
 class Attribute implements RenderItem
 {
