+++ src/PhpDia/Tests/Dia/ElementTest.php
@@ -2,9 +2,9 @@
 
 namespace PhpDia\Tests\Dia;
 
-use PhpDia\Dia\Attribute;
-use PhpDia\Dia\ClassElement;
-use PhpDia\Dia\Operation;
+use PhpDia\Dia\Xml\Attribute;
+use PhpDia\Dia\Xml\ClassElement;
+use PhpDia\Dia\Xml\Operation;
 use PhpDia\Dia\Values\BoundingBox;
 use PhpDia\Dia\Values\Position;
 use PHPUnit\Framework\TestCase;
